requirements.txt
streamlit
langgraph
chromadb
sentence-transformers
groq
pypdf
python-pptx
ragas
datasets